Move external dependency to register time.
